Move package to quickstart
