--- vhpidirect/linking/package/cSharedVar.h
@@ -1,6 +1,0 @@
-#include<stdio.h>
-
-int var;
-
-int* getInt_ptr();
-void printInt();
