+++ vhpidirect/quickstart/package/cSharedVar.h
@@ -1,0 +1,6 @@
+#include<stdio.h>
+
+int var;
+
+int* getInt_ptr();
+void printInt();
